+++ packages/core/command-data-types/src/types/Condition.ts
@@ -7,6 +7,7 @@ export interface EdgeCondition {
 
 export interface EdgeConditionAssertion {
 	setpoint: {
+		id: string;
 		value: string;
 	}
 	variable: {