+++ packages/app/hivecommand-api/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.52](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.51...v1.4.15-alpha.52) (2022-05-27)
+
+**Note:** Version bump only for package @hive-command/api
+
+
+
+
+
+## [1.4.15-alpha.37](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.36...v1.4.15-alpha.37) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/api
+
+
+
+
+
+## [1.4.15-alpha.36](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.35...v1.4.15-alpha.36) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/api
+
+
+
+
+
 ## [1.4.15-alpha.26](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.25...v1.4.15-alpha.26) (2022-05-22)
 
 **Note:** Version bump only for package @hive-command/api